Started to rework the standard PDF font data.
Instead of hardcoding things as C array, the afm files and encoding files
are now included.

As a consequence it is a bit less user friendly to use the standard fonts
since they will have to be loaded before.

The loading mechanism is not yet done. But the API has already moved a
little towards that goal.